--- a/crates/database/src/orders.rs
+++ b/crates/database/src/orders.rs
@@ -500,7 +500,7 @@ pub fn user_orders<'a>(
 " WHERE o.owner = $1",
 " ORDER BY creation_timestamp DESC LIMIT $2 * ($3 + 1) ) ",
 " UNION ",
-" (SElECT ", ORDERS_SELECT,
+" (SELECT ", ORDERS_SELECT,
 " FROM ", ORDERS_FROM,
 " LEFT OUTER JOIN onchain_placed_orders onchain_o on onchain_o.uid = o.uid",
 " WHERE onchain_o.sender = $1 ",