--- a/database/sql/V041__create_index_on_orders_with_owner_and_timestamp.sql
+++ b/database/sql/V041__create_index_on_orders_with_owner_and_timestamp.sql
@@ -1,3 +1,4 @@
--- The index can be traversed in both directions anyway but we usually fetch user orders with this
--- ordering so using it here too.
+-- To optimize the performance of the user_orders query, we introduce a new index that allows
+-- us to quickly get the latest orders from a owner
 CREATE INDEX user_order_creation_timestamp ON orders USING BTREE (owner, creation_timestamp DESC);
+