@nlordell 2022-11-11
@josojo - would it be possible to understand what PostgreSQL is doing with these queries using `EXPLAIN ANALYZE` on the query?