@josojo 2022-11-11
@nlordell I am not sure what you wanna see from there.

I put the output here
https://docs.google.com/spreadsheets/d/1U15MfFvIQbKmnNP2KA9FB-XuEkbki_IfC4O_K9G7i1o/edit#gid=1863284260
And I compare the difference with and without limit in the subqueries. I highlighted the difference in green. But I am not sure whether this gives me more confidence.

I will also create one more benchmark between this PR and prod code with the benchmark tools.

But I did not make any progress with [duplicating the db](https://cowservices.slack.com/archives/C6Z2XNL5Q/p1667989826720819). Anuj support would be needed for that.